+++ extensions/iceworks-component-builder/src/extension.ts
@@ -28,7 +28,7 @@ export function activate(context: vscode.ExtensionContext) {
   initExtensionConfiguration(globalState);
 
   function activeWebview() {
-    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', '生成组件', ViewColumn.One, {
+    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', '生成组件 - Iceworks', ViewColumn.One, {
       enableScripts: true,
       retainContextWhenHidden: true,
     });