+++ extensions/iceworks-project-creator/web/src/constant.ts
@@ -1,0 +1,16 @@
+
+export const mainScaffoldsList = [
+  '@alifd/fusion-design-pro',
+  '@alifd/scaffold-lite',
+  '@alifd/scaffold-simple'
+]
+export const tsScaffoldsList = [
+  ...mainScaffoldsList,
+  '@icedesign/stark-layout-scaffold',
+  '@icedesign/stark-child-scaffold'
+]
+
+export const jsScaffoldsList = [
+  '@alifd/fusion-design-pro-js',
+  '@alifd/scaffold-lite-js'
+]